Vastly improve landscape orientation UI.
This is a big change. Previously phone orientation changes would cause the UI to autolayout. This put the shutter button in an awkware place to reach. And many of the screens didn't look good in landscape mode.

To remedy this, the app is now fixed to the main portrait orientation. Rotating the phone leaves all elements in place. Rather than adjusting the entire UI, there are a select view elements that will stay in place but rotate their orientation with little animations (the buttons on the camera and photos screen, and the flash settings dialog). In addition, the photo review will adjust itself.

Fixed up lots of other UI jankiness on the camera and photo view screen - it doesn't jerk around anymore.

Fixes #5
Fixes #22
Fixes #58
Fixes #97