Camera preview should show entire photo.
The preview view was previously slightly zoomed so the entire screen would show an image. It looked better but also meant part of the image was cropped so the final photo would have a different result.

This is no longer the case - the photo on screen matches the final photo. A side effect is there are now black bars at the top/bottom but this is ok.

Fixes #109