Photo quality is about a zillion times better.
Now when a photo is taken, between when the light comes on and the photo is captured, the app pauses allowing the camera hardware time to finish adjusting focus/exposure/whitebalance. When all of these have settled down it will take the shot.

Tests are showing photos even in pitch black scenes are now coming out very high quality. Happy with the result :)